src/infra/repository/postgresql/user-profile.ts
import { PostgreSQLUser } from '@/infra/repository/postgresql/entities'
import { PgRepository } from '@/infra/repository/postgresql/repository'
import { LoadUserProfile, SaveUserPicture } from '@/domain/contracts/repository'

export class PgUserProfileRepository extends PgRepository implements SaveUserPicture, LoadUserProfile {
  async savePicture ({ id, pictureUrl, initials }: SaveUserPicture.Input): Promise<void> {
    const pgUserRepo = this.getRepository(PostgreSQLUser)
    await pgUserRepo.update({ id: parseInt(id) }, { pictureUrl, initials })
  }

  async load ({ id }: LoadUserProfile.Input): Promise<LoadUserProfile.Output> {
    const pgUserRepo = this.getRepository(PostgreSQLUser)
    const pgUser = await pgUserRepo.findOne({ id: parseInt(id) })
    if (pgUser !== undefined) return { name: pgUser.name ?? undefined }
  }
}
